requirements.txt
dhanhq>=2.0.2
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
pytz>=2024.1
